feat: extracted ECDSA key pair loader
